--- src/actions/wit.actions.js
@@ -1,9 +1,0 @@
-// import Axios from 'axios';
-
-// const wit = '/api/wit/query';
-
-// export const WIT_QUERY = 'WIT_QUERY';
-// // const WIT_URL = 'https://api.wit.ai/message';
-
-// export const witQuery = (witQuery) => {
-//   return(dispatch) => {}
