+++ src/index.css
@@ -3,3 +3,100 @@ body {
   margin: 0;
   padding: 0;
 }
+
+.App {
+  background-color: pink;
+}
+
+
+
+.App-header {
+  background-color: black;
+  position: relative;
+}
+  .App-header img {
+    border: 4px solid cyan;
+    border-radius: 10px;
+    margin: 15px;
+  }
+
+
+
+.App-bar {
+  border: 3px dashed white;
+  display: inline-block;
+  position: absolute;
+  right: 15px;
+  bottom: 15px;
+}
+  .App-bar ul {
+    display: table;
+    list-style-type: none;
+    margin-left: 0;
+    padding-left: 0;
+  }
+  .App-bar li {
+    border: 1px solid magenta;
+    padding: 5px;
+    margin: 5px;
+    display: inline-block;
+  }
+  .App-bar li a {
+    border: 1px solid cyan;
+    padding: 5px;
+    margin: 10px;
+    font-size: 1.5em;
+    font-family: sans-serif;
+    color: yellow;
+  }
+  .App-bar li a:hover {
+    text-decoration: underline;
+    cursor: pointer;
+  }
+
+
+.App-chat {
+  border: 3px dashed magenta;
+}
+
+.App-chat-window {
+  border: 1px solid black;
+  font-size: 2em;
+  width: 80%;
+  margin: 0 auto;
+}
+  .App-chat-window span {
+    display: block;
+    border: 1px solid white;
+    font-family: sans-serif;
+  }
+
+
+
+
+.App-input {
+  border: 1px dashed black;
+  height: 100%;
+  position: relative;
+}
+  .App-input div {
+    width: 50%;
+    margin: 0 auto;
+    border: 1px solid black;
+  }
+  .App-input textarea {
+    border: 1px solid magenta;
+    margin: 5px;
+    display: block;
+    width: 75%;
+  }
+  .App-input button {
+    border: 1px solid yellow;
+    border-radius: 5px;
+    background-color: yellow;
+    font-weight: bold;
+    margin: 5px;
+  }
+  .App-input button:hover {
+    cursor: pointer;
+  }
