+++ src/index.js
@@ -1,19 +1,15 @@
 /*REACT ENTRY POINT*/
-/*REACT ENTRY POINT*/
-/*REACT ENTRY POINT*/
-
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
 import { createStore, applyMiddleware } from 'redux';
-import { connect } from 'react-redux';
 
 import reducers from './reducers';
 import thunk from 'redux-thunk';
 import registerServiceWorker from './lib/registerServiceWorker';
 
 /*COMPONENTS*/
-import App from './containers/App/App';
+import App from './containers/App/';
 
 import './index.css';
