+++ src/reducers/index.js
@@ -1,13 +1,8 @@
 import { combineReducers } from 'redux';
-import TopicsReducer from './reducer_topic';
-import ActiveTopic from './reducer_active_topic';
-import NigelReducer from './reducer_nigel_topics';
-
-const initialState = {};
+import topicsList from './topics.reducer';
 
 const rootReducer = combineReducers({
-  topics: TopicsReducer,
-  ActiveTopic: ActiveTopic
+  topicsList : topicsList
 });
 
 export default rootReducer;
