--- src/reducers/reducer_nigel_topics.js
@@ -1,1 +1,0 @@
-// reduver for nigel db
